Move subscription forms to `subscription/forms.py`
